backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pymongo
python-dotenv
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
pydantic[email]==2.10.0
python-jose[cryptography]
email-validator